fix: remove reflectivity cast to uint32_t
since the reflectivity is always an array of 8 bits, it is uncessary to cast it to 32 bits

+++ tasks/Task.hpp
@@ -122,7 +122,7 @@ argument.
         ouster::LidarScan acquireData();
         void convertDataAndWriteOutput(ouster::LidarScan& scan);
         void writeIMUSample(std::unique_ptr<uint8_t[]> const& pkt_buffer);
-        ouster::img_t<uint32_t> getReflectivity(ouster::LidarScan const& scan);
+        ouster::img_t<uint8_t> getReflectivity(ouster::LidarScan const& scan);
     };
 }
